requirements.txt
asgiref==3.3.1
Django==3.1.4
django-3-jet==1.0.8
django-model-utils==4.1.1
django-notifications-hq==1.6.0
jsonfield==3.1.0
Pillow
python-slugify==4.0.1
pytz==2020.5
six==1.15.0
slugify==0.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
text-unidecode==1.3
